migrate ConditionVariable
