+++ src/main/java/org/betonquest/betonquest/quest/registry/CoreQuestTypes.java
@@ -144,7 +144,8 @@ import org.betonquest.betonquest.quest.event.time.TimeEventFactory;
 import org.betonquest.betonquest.quest.event.velocity.VelocityEventFactory;
 import org.betonquest.betonquest.quest.event.weather.WeatherEventFactory;
 import org.betonquest.betonquest.quest.registry.type.ConditionTypeRegistry;
-import org.betonquest.betonquest.variables.ConditionVariable;
+import org.betonquest.betonquest.quest.registry.type.VariableTypeRegistry;
+import org.betonquest.betonquest.quest.variable.condition.ConditionVariableFactory;
 import org.betonquest.betonquest.variables.GlobalPointVariable;
 import org.betonquest.betonquest.variables.GlobalTagVariable;
 import org.betonquest.betonquest.variables.ItemDurabilityVariable;
@@ -220,7 +221,7 @@ public class CoreQuestTypes {
         registerConditions(questTypeRegistries.getConditionTypes());
         registerEvents();
         registerObjectives();
-        registerVariables();
+        registerVariables(questTypeRegistries.getVariableTypes());
     }
 
     private void registerConditions(final ConditionTypeRegistry conditionTypes) {
@@ -373,8 +374,8 @@ public class CoreQuestTypes {
         }
     }
 
-    private void registerVariables() {
-        betonQuest.registerVariable("condition", ConditionVariable.class);
+    private void registerVariables(final VariableTypeRegistry variables) {
+        variables.register("condition", new ConditionVariableFactory());
         betonQuest.registerVariable("globalpoint", GlobalPointVariable.class);
         betonQuest.registerVariable("globaltag", GlobalTagVariable.class);
         betonQuest.registerVariable("item", ItemVariable.class);
