+++ src/main/java/org/betonquest/betonquest/quest/variable/condition/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.variable.Variable Variable} implementation for the condition variable.
+ */
+package org.betonquest.betonquest.quest.variable.condition;
